pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "move-faults-monorepo"
version = "0.1.0"
description = "Monorepo for the Philippine Open Geodesy Framework (POGF) and related tools."
requires-python = ">=3.11"
license = "MIT"
authors = [
    { name = "Alfie Pelicano", email = "alfieprojects.dev@gmail.com" },
]

# This section should remain empty; dependencies are defined per-project below.
dependencies = []

[project.scripts]
# Define command-line entry points for all tools in the monorepo
drive-archaeologist = "drive_archaeologist.cli:main"
drive-arch = "drive_archaeologist.cli:main"
vadase-ingestor = "scripts.run_ingestor:main"
vadase-validate = "scripts.validate_parser:main"

[project.optional-dependencies]
# Dependencies for the drive-archaeologist tool
drive-archaeologist = [
    "click>=8.1.0",
    "rich>=13.0.0",
    "pathvalidate>=3.0.0",
    "filetype>=1.2.0",
    "pillow>=10.0.0",
    "py7zr>=0.20.0",
    "rarfile>=4.0",
    "pandas>=2.0.0",
    "pytesseract>=0.3.10",
    "opencv-python>=4.8.0",
    "pypdf>=3.0.0",
    "watchdog>=3.0.0",
    "plotly>=5.0.0",
    "jinja2>=3.0.0",
    "whoosh>=2.7.4",
]

# Dependencies for the vadase-rt-monitor service
vadase-rt-monitor = [
    "asyncpg>=0.29.0",
    "python-dotenv>=1.0.0",
    "pyyaml>=6.0.1",
    "pydantic>=2.5.3",
    "psycopg2-binary>=2.9.9",
    "aiosmtplib>=3.0.1",
    "python-telegram-bot>=20.7",
    "structlog>=24.1.0",
    "prometheus-client>=0.19.0",
    "httpx>=0.28.1",
]

# Shared development dependencies
dev = [
    "pytest>=7.4.3",
    "pytest-cov>=4.1.0",
    "pytest-asyncio>=0.21.1",
    "black>=23.12.1",
    "ruff>=0.1.9",
    "mypy>=1.8.0",
    "jupyterlab>=4.0.10",
    "matplotlib>=3.8.2",
    "pandas>=2.1.4",
    "seaborn>=0.13.0",
    "ipython>=8.12.0",
    "uv",
]

[tool.ruff]
line-length = 100
target-version = "py311"
select = ["E", "F", "I", "B", "C4", "UP"]
ignore = ["E501", "B008"]

[tool.ruff.per-file-ignores]
"__init__.py" = ["F401"]

[tool.mypy]
python_version = "3.11"
warn_return_any = true
warn_unused_configs = true
ignore_missing_imports = true

[tool.black]
line-length = 100
target-version = ['py311']
